+++ docs/changelog.md
@@ -1,6 +1,8 @@
 ## Changelog
 
 ## 0.5.7 - Unreleased
+- Updated dependencies
+- Fixed mkdocstrings
 
 ## 0.5.6 - 2023-04-28
 - Don´t install tika more than once per session